Fixing the filter search button

+++ sync-hub-be/index.js
@@ -101,64 +101,10 @@ app.post("/send-messages", async (req, res) => {
   // message is from: resEmail(sender) => to: reqEmail(receiver)
   console.log(message, reqEmail, resEmail);
 
-  // check if sender of message has a chat history
-  let senderChats = await ChatHistory.findOne({ owner: resEmail });
-
   // chats array to store the chat history between sender and receiver of message that will be sent back to the front end
   let chatsArray = [];
 
-  // if sender doesn't have a chat history
-  if (!senderChats) {
-    let chatHistory = {
-      owner: resEmail,
-      chats: [
-        {
-          with: reqEmail,
-          history: [
-            {
-              id: resEmail,
-              text: message,
-            },
-          ],
-        },
-      ],
-    };
-    //create a chat history for sender and update it with the incoming message details
-    let newChatHistory = await ChatHistory.create(chatHistory);
-    // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
-    chatsArray = [{ id: resEmail, text: message }];
-  }
-  // if sender has a chat history
-  else if (senderChats && senderChats.owner === resEmail) {
-    let flag = "";
-    // check if any chat history exists between sender and receiver of message
-    for (let chats of senderChats.chats) {
-      // if any chat history exists between sender and receiver of message
-      if (chats.with === reqEmail) {
-        let newChats = [...chats.history];
-        // update chat history with the incoming message details
-        newChats.push({ id: resEmail, text: message });
-        chats.history = newChats;
-        // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
-        chatsArray = newChats;
-        flag = "on";
-      }
-    }
-    // if no chat history exists between sender and receiver of message
-    if (flag === "") {
-      // create a new chat history between sender and receiver of message
-      let newChats = {
-        with: reqEmail,
-        history: [{ id: resEmail, text: message }],
-      };
-      // update chat history with the incoming message details
-      senderChats.chats.push(newChats);
-      // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
-      chatsArray = [{ id: resEmail, text: message }];
-    }
-    // save updated details to sender chat history database
-    senderChats.save();
-  }
+ 
 
   //check if receiver of message has a chat history
   let receiverChats = await ChatHistory.findOne({ owner: reqEmail });
@@ -216,6 +162,62 @@ app.post("/send-messages", async (req, res) => {
     receiverChats.save();
   }
 
+   // check if sender of message has a chat history
+   let senderChats = await ChatHistory.findOne({ owner: resEmail });
+
+   // if sender doesn't have a chat history
+   if (!senderChats) {
+     let chatHistory = {
+       owner: resEmail,
+       chats: [
+         {
+           with: reqEmail,
+           history: [
+             {
+               id: resEmail,
+               text: message,
+             },
+           ],
+         },
+       ],
+     };
+     //create a chat history for sender and update it with the incoming message details
+     let newChatHistory = await ChatHistory.create(chatHistory);
+     // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
+     chatsArray = [{ id: resEmail, text: message }];
+   }
+   // if sender has a chat history
+   else if (senderChats && senderChats.owner === resEmail) {
+     let flag = "";
+     // check if any chat history exists between sender and receiver of message
+     for (let chats of senderChats.chats) {
+       // if any chat history exists between sender and receiver of message
+       if (chats.with === reqEmail) {
+         let newChats = [...chats.history];
+         // update chat history with the incoming message details
+         newChats.push({ id: resEmail, text: message });
+         chats.history = newChats;
+         // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
+         chatsArray = newChats;
+         flag = "on";
+       }
+     }
+     // if no chat history exists between sender and receiver of message
+     if (flag === "") {
+       // create a new chat history between sender and receiver of message
+       let newChats = {
+         with: reqEmail,
+         history: [{ id: resEmail, text: message }],
+       };
+       // update chat history with the incoming message details
+       senderChats.chats.push(newChats);
+       // update chats array that stores chat history between sender and receiver of message that will later be sent back to the front end
+       chatsArray = [{ id: resEmail, text: message }];
+     }
+     // save updated details to sender chat history database
+     senderChats.save();
+   }
+
   // console.log("EMIT");
   // send chats array, receiver of message sent, sender of message sent as an array to the front end
   // io.emit("receive", { message: [chatsArray, reqEmail, resEmail] });
@@ -263,7 +265,7 @@ app.post("/get-user-chats", async (req, res) => {
 app.post("/signup", async (req, res) => {
   console.log(req.body);
   try {
-    let user = await Users.find({ email: req.body.email.trim() });
+    let user = await Users.find({ email: req.body.email });
     // console.log(1)
     if (user.length >= 1) {
       res.status(202).send({ message: "User Already Exists" });
@@ -377,15 +379,11 @@ app.put("/change-user-details", async (req, res) => {
 
 // DELETE METHODS
 app.delete("/user/delete-account", async (req, res) => {
-  // console.log(req.body);
+  console.log(req.body);
   try {
     let user = await Users.findOneAndDelete({ email: req.body.email });
-    await ChatHistory.findOneAndDelete({
-      owner: req.body.email,
-    });
-    await Requests.findOneAndDelete({
-      reqShaker: req.body.email,
-    });
+    await ChatHistory.findOneAndDelete({ owner: req.body.email });
+    await Requests.findOneAndDelete({ reqShaker: req.body.email });
     if (user) {
       res.status(201).send({ message: "Account deleted successfully!" });
     } else {
